refactor(competence): fold insights coverage spec builders into the insightsCycle component (CA-65)
