+++ packages/competence/bin/static/index.html
@@ -13,7 +13,6 @@
     <link href="https://fonts.googleapis.com/css2?family=JetBrains+Mono:wght@400;500;600&display=swap" rel="stylesheet">
     <script defer src="/static/scripts/ti-framework.js" nonce='{ti-nonce-placeholder}'></script>
     <script defer src="/static/scripts/ti-charts.js" nonce='{ti-nonce-placeholder}'></script>
-    <script defer src="/static/scripts/insights-cycle-specs.js" nonce='{ti-nonce-placeholder}'></script>
     <script defer src="/static/scripts/competence-user-interface.js" nonce='{ti-nonce-placeholder}'></script>
     <script defer src="/static/scripts/lib/alpinejs-csp.min.js" nonce='{ti-nonce-placeholder}'></script>
     <script defer src="/static/scripts/lib/htmx.min.js" nonce='{ti-nonce-placeholder}'></script>
